+++ library/core/src/main/java/com/google/android/exoplayer2/PlayerMessage.java
@@ -16,8 +16,8 @@
 package com.google.android.exoplayer2;
 
 import android.os.Handler;
+import android.os.Looper;
 import androidx.annotation.Nullable;
-import androidx.annotation.VisibleForTesting;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Clock;
 import java.util.concurrent.TimeoutException;
@@ -55,11 +55,12 @@ public final class PlayerMessage {
 
   private final Target target;
   private final Sender sender;
+  private final Clock clock;
   private final Timeline timeline;
 
   private int type;
   @Nullable private Object payload;
-  private Handler handler;
+  private Looper looper;
   private int windowIndex;
   private long positionMs;
   private boolean deleteAfterDelivery;
@@ -77,7 +78,8 @@ public final class PlayerMessage {
    *     set to {@link Timeline#EMPTY}, any position can be specified.
    * @param defaultWindowIndex The default window index in the {@code timeline} when no other window
    *     index is specified.
-   * @param defaultHandler The default handler to send the message on when no other handler is
+   * @param clock The {@link Clock}.
+   * @param defaultLooper The default {@link Looper} to send the message on when no other looper is
    *     specified.
    */
   public PlayerMessage(
@@ -85,11 +87,13 @@ public final class PlayerMessage {
       Target target,
       Timeline timeline,
       int defaultWindowIndex,
-      Handler defaultHandler) {
+      Clock clock,
+      Looper defaultLooper) {
     this.sender = sender;
     this.target = target;
     this.timeline = timeline;
-    this.handler = defaultHandler;
+    this.looper = defaultLooper;
+    this.clock = clock;
     this.windowIndex = defaultWindowIndex;
     this.positionMs = C.TIME_UNSET;
     this.deleteAfterDelivery = true;
@@ -142,22 +146,28 @@ public final class PlayerMessage {
     return payload;
   }
 
+  /** @deprecated Use {@link #setLooper(Looper)} instead. */
+  @Deprecated
+  public PlayerMessage setHandler(Handler handler) {
+    return setLooper(handler.getLooper());
+  }
+
   /**
-   * Sets the handler the message is delivered on.
+   * Sets the {@link Looper} the message is delivered on.
    *
-   * @param handler A {@link Handler}.
+   * @param looper A {@link Looper}.
    * @return This message.
    * @throws IllegalStateException If {@link #send()} has already been called.
    */
-  public PlayerMessage setHandler(Handler handler) {
+  public PlayerMessage setLooper(Looper looper) {
     Assertions.checkState(!isSent);
-    this.handler = handler;
+    this.looper = looper;
     return this;
   }
 
-  /** Returns the handler the message is delivered on. */
-  public Handler getHandler() {
-    return handler;
+  /** Returns the {@link Looper} the message is delivered on. */
+  public Looper getLooper() {
+    return looper;
   }
 
   /**
@@ -287,19 +297,19 @@ public final class PlayerMessage {
    * Blocks until after the message has been delivered or the player is no longer able to deliver
    * the message.
    *
-   * <p>Note that this method can't be called if the current thread is the same thread used by the
-   * message handler set with {@link #setHandler(Handler)} as it would cause a deadlock.
+   * <p>Note that this method must not be called if the current thread is the same thread used by
+   * the message {@link #getLooper() looper} as it would cause a deadlock.
    *
    * @return Whether the message was delivered successfully.
    * @throws IllegalStateException If this method is called before {@link #send()}.
    * @throws IllegalStateException If this method is called on the same thread used by the message
-   *     handler set with {@link #setHandler(Handler)}.
+   *     {@link #getLooper() looper}.
    * @throws InterruptedException If the current thread is interrupted while waiting for the message
    *     to be delivered.
    */
   public synchronized boolean blockUntilDelivered() throws InterruptedException {
     Assertions.checkState(isSent);
-    Assertions.checkState(handler.getLooper().getThread() != Thread.currentThread());
+    Assertions.checkState(looper.getThread() != Thread.currentThread());
     while (!isProcessed) {
       wait();
     }
@@ -310,14 +320,14 @@ public final class PlayerMessage {
    * Blocks until after the message has been delivered or the player is no longer able to deliver
    * the message or the specified timeout elapsed.
    *
-   * <p>Note that this method can't be called if the current thread is the same thread used by the
-   * message handler set with {@link #setHandler(Handler)} as it would cause a deadlock.
+   * <p>Note that this method must not be called if the current thread is the same thread used by
+   * the message {@link #getLooper() looper} as it would cause a deadlock.
    *
    * @param timeoutMs The timeout in milliseconds.
    * @return Whether the message was delivered successfully.
    * @throws IllegalStateException If this method is called before {@link #send()}.
    * @throws IllegalStateException If this method is called on the same thread used by the message
-   *     handler set with {@link #setHandler(Handler)}.
+   *     {@link #getLooper() looper}.
    * @throws TimeoutException If the {@code timeoutMs} elapsed and this message has not been
    *     delivered and the player is still able to deliver the message.
    * @throws InterruptedException If the current thread is interrupted while waiting for the message
@@ -325,14 +335,8 @@ public final class PlayerMessage {
    */
   public synchronized boolean blockUntilDelivered(long timeoutMs)
       throws InterruptedException, TimeoutException {
-    return blockUntilDelivered(timeoutMs, Clock.DEFAULT);
-  }
-
-  @VisibleForTesting()
-  /* package */ synchronized boolean blockUntilDelivered(long timeoutMs, Clock clock)
-      throws InterruptedException, TimeoutException {
     Assertions.checkState(isSent);
-    Assertions.checkState(handler.getLooper().getThread() != Thread.currentThread());
+    Assertions.checkState(looper.getThread() != Thread.currentThread());
 
     long deadlineMs = clock.elapsedRealtime() + timeoutMs;
     long remainingMs = timeoutMs;
@@ -340,11 +344,9 @@ public final class PlayerMessage {
       wait(remainingMs);
       remainingMs = deadlineMs - clock.elapsedRealtime();
     }
-
     if (!isProcessed) {
       throw new TimeoutException("Message delivery timed out.");
     }
-
     return isDelivered;
   }
 }